Extract custom testem runners to `lib/testem/` subfolder
